Started to implement some of the node logic
Implemented almost all of the node registration logic on the API

+++ lib/keys/generate_test.go
@@ -2,12 +2,15 @@ package keys
 
 import (
 	"github.com/stretchr/testify/assert"
-	"../config"
+	"golang.org/x/crypto/ssh"
 	"io/ioutil"
-	"../util"
 	"testing"
+	"reflect"
 	"log"
 	"os"
+
+	"../config"
+	"../util"
 )
 
 func TestMain(m *testing.M) {
@@ -22,9 +25,28 @@ func TestMain(m *testing.M) {
 }
 
 func TestGenerate(t *testing.T) {
+	privKey, pubKey := Generate()
+
+	privateKey, err := ssh.ParseRawPrivateKey([]byte(privKey))
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	signer, err := ssh.NewSignerFromKey(privateKey)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	publicKey := getPubKey(t, pubKey)
+
+	assert.True(t, reflect.DeepEqual(signer.PublicKey(), publicKey))
+}
+
+func TestWriteToFile(t *testing.T) {
 	conf := config.GetConfig()
 
-	Generate()
+	privKey, pubKey := Generate()
+	WriteToFile(privKey, pubKey)
 
 	publicFile := conf.GetString("id_rsa_pub_path")
 	privateFile := conf.GetString("id_rsa_path")
@@ -37,12 +59,14 @@ func TestGenerateOverwrite(t *testing.T) {
 	conf := config.GetConfig()
 	privateFile := conf.GetString("id_rsa_path")
 
-	Generate()
+	privKey, pubKey := Generate()
+	WriteToFile(privKey, pubKey)
 
 	dat1, err := ioutil.ReadFile(privateFile)
 	check(err)
 
-	Generate()
+	privKey, pubKey = Generate()
+	WriteToFile(privKey, pubKey)
 
 	dat2, err := ioutil.ReadFile(privateFile)
 	check(err)
@@ -52,8 +76,10 @@ func TestGenerateOverwrite(t *testing.T) {
 
 func removeFiles(files ...string) {
 	for _, file := range files {
-		err := os.Remove(file)
-		check(err)
+		if util.FileExists(file) {
+			err := os.Remove(file)
+			check(err)
+		}
 	}
 }
 
@@ -61,4 +87,12 @@ func check(e error) {
     if e != nil {
         panic(e)
     }
+}
+
+func getPubKey(t *testing.T, pubKey string) ssh.PublicKey {
+	parsedKey, _, _, _, err := ssh.ParseAuthorizedKey([]byte(pubKey))
+    if err != nil {
+        t.Errorf("ERROR! %s", err)
+	}
+	return parsedKey
 }
